api-review-agent/requirements.txt
langchain>=0.1.0
openai>=1.0.0
PyGithub>=2.1.0
atlassian-python-api>=3.40.0
pyyaml>=6.0
jinja2>=3.1.0
python-dotenv>=1.0.0
